+++ plugin/src/test/java/com/stratio/cassandra/lucene/schema/SchemaTest.java
@@ -31,6 +31,7 @@ import org.apache.cassandra.exceptions.InvalidRequestException;
 import org.apache.cassandra.thrift.CfDef;
 import org.apache.cassandra.thrift.ColumnDef;
 import org.apache.cassandra.thrift.IndexType;
+import org.apache.cassandra.thrift.ThriftConversion;
 import org.apache.cassandra.utils.ByteBufferUtil;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.analysis.en.EnglishAnalyzer;
@@ -113,7 +114,7 @@ public class SchemaTest {
                                  .setColumn_metadata(columnDefinitions)
                                  .setKeyspace("Keyspace1")
                                  .setName("Standard1");
-        CFMetaData metadata = CFMetaData.fromThrift(cfDef);
+        CFMetaData metadata = ThriftConversion.fromThrift(cfDef);
 
         Schema schema = SchemaBuilders.schema().mapper("field1", stringMapper()).mapper("field2", textMapper()).build();
         schema.validate(metadata);